% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{as_internal_format}
\alias{as_internal_format}
\alias{as_internal_format.ospar}
\alias{as_internal_format.wide}
\alias{as_internal_format.long}
\title{Signature from Files}
\usage{
as_internal_format(x, ...)

\method{as_internal_format}{ospar}(x, ...)

\method{as_internal_format}{wide}(x, ...)

\method{as_internal_format}{long}(x, ...)
}
\arguments{
\item{x}{object of which the signature has to be determined}

\item{\dots}{further arguments passed to or from other methods.}
}
\description{
Generic function for retrieving the file signature
}
\section{Methods (by class)}{
\itemize{
\item \code{ospar}: convert OSPAR data to internal representation.

\item \code{wide}: convert wide data to internal representation.

\item \code{long}: convert long data to internal representation.
}}

