% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{attach_dep}
\alias{attach_dep}
\title{Attach Dependencies}
\usage{
attach_dep(tag, script)
}
\arguments{
\item{tag}{Tag to attach the dependencies to.}
}
\description{
Attach dependencies to a tag.
}
\keyword{internal}
