% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui.R
\name{create_litter_project}
\alias{create_litter_project}
\title{Create Project Directory}
\usage{
create_litter_project(path = NULL)
}
\arguments{
\item{path}{(Existing) directory name}
}
\description{
Fills an empty directory (\code{path}) with example files. If the
\code{path}' argument is missing or \code{NULL}, 
a Tcl/Tk dialogue will be started.
}
