% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{cv}
\alias{cv}
\title{Coefficient of Variation}
\usage{
cv(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector}

\item{na.rm}{logical. Should missing values be removed?}
}
\value{
coefficient of variation (numeric vector of length 1).
}
\description{
Coefficient of Variation
}
\references{
\url{https://en.wikipedia.org/wiki/Coefficient_of_variation}
}
