% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{enumerate.character}
\alias{enumerate.character}
\title{Enumerate Character Vector}
\usage{
\method{enumerate}{character}(x, ...)
}
\arguments{
\item{x}{character vector}

\item{\dots}{further arguments passed to or from other methods.}
}
\value{
\code{character} vector of length 1, with elements
separated by a comma except for the last element which
is prepended by "and".
}
\description{
Collapsing a \code{character} vector of length n, to a
\code{character} vector of length 1.
}
\examples{
enumerate("apples")
enumerate(c("apples", "oranges"))
enumerate(c("apples", "oranges", "pears"))
}
