% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{enumerate}
\alias{enumerate}
\title{Enumerate Objects}
\usage{
enumerate(x, ...)
}
\arguments{
\item{x}{object to enumerate}

\item{\dots}{further arguments passed to or from other methods.}
}
\description{
Generic function for enumerating objects
}
\seealso{
\code{\link{enumerate.character}}
}
