% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{enumerate.sequenized}
\alias{enumerate.sequenized}
\title{Convert Sequenized Output to Character String}
\usage{
\method{enumerate}{sequenized}(x, ...)
}
\arguments{
\item{x}{object of class \code{sequenized}.}

\item{\dots}{further arguments passed to or from other methods.}
}
\value{
string representation (character vector of length 1) of
a sequenized object
}
\description{
Convert Sequenized Output to Character String
}
\seealso{
\code{\link{sequenize.integer}}
}
