% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_write_access}
\alias{has_write_access}
\title{Check Write Permission}
\usage{
has_write_access(path)
}
\arguments{
\item{path}{filename}
}
\value{
\code{TRUE} if write access, \code{FALSE} if not
}
\description{
Simple wrapper for \code{\link{file.access}} with \code{mode=2}
}
