% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{intercept}
\alias{intercept}
\title{Intercept}
\usage{
intercept(x, ...)
}
\arguments{
\item{x}{object}

\item{\dots}{further arguments passed to or from other methods.}
}
\value{
estimate of the intercept (numeric vector of length 1).
}
\description{
Extract the intercept from object \code{x}.
}
