% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{iod}
\alias{iod}
\title{Index of Dispersion}
\usage{
iod(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector}

\item{na.rm}{logical. Should missing values be removed?}
}
\value{
index of dispersion (numeric vector of length 1).
}
\description{
A normalized measure of the dispersion of a probability distribution.
}
\references{
\url{https://en.wikipedia.org/wiki/Index_of_dispersion}
}
