% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_date_format}
\alias{is_date_format}
\title{Check Date Format}
\usage{
is_date_format(x, format)
}
\arguments{
\item{x}{object of class \code{character} or \code{Date}}

\item{format}{required date format (see \code{\link{strptime}})}
}
\value{
\code{TRUE} if \code{x} complies with \code{format}, 
and \code{FALSE} otherwise.
}
\description{
Checks if the data format \code{x} complies with \code{format}.
}
\examples{
is_date_format("2019-05-14", "\%Y-\%m-\%d")
}
