% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_group_code}
\alias{is_group_code}
\title{Test For Group Code}
\usage{
is_group_code(x)
}
\arguments{
\item{x}{\code{character} vector to test}
}
\value{
\code{TRUE} if test passes, \code{FALSE} otherwise.
}
\description{
Test if \code{x} contains a valid litter group code.
}
