% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_type_code}
\alias{is_type_code}
\title{Test For Type Code}
\usage{
is_type_code(x)
}
\arguments{
\item{x}{\code{character} vector to test}
}
\value{
\code{TRUE} if test passes, \code{FALSE} otherwise.
}
\description{
Test if \code{x} contains a valid litter type code.
}
