% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{kendall_var_s}
\alias{kendall_var_s}
\title{Mann-Kendall Variance of S Statistic}
\usage{
kendall_var_s(x, t = seq_along(x))
}
\arguments{
\item{x}{observations}

\item{t}{time index}
}
\description{
Mann-Kendall Variance of S Statistic
}
\references{
Gilbert, R.O., 1987. Statistical Methods for Environmental
Pollution Monitoring.

Van Belle and Hughes, 1984, Nonparametric Tests for Trend in 
Water Quality. Water Resources Research 20:127-136
}
