% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/els.R
\name{lActionButton}
\alias{lActionButton}
\title{Action Button}
\usage{
lActionButton(name, label = "Click me", ..., id = NULL, class = NULL)
}
\arguments{
\item{name}{Name of the input.}

\item{label}{Label of the button.}

\item{...}{Passed to props.}

\item{id}{Id of the input.}

\item{class}{Class of the button.}
}
\description{
Create an action button.
}
