% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-button.R
\name{lit_action_button_update}
\alias{lit_action_button_update}
\title{Update Action Button}
\usage{
lit_action_button_update(session, selector, ..., label = NULL)
}
\arguments{
\item{session}{A valid shiny session.}

\item{selector}{A selector for the buttons to update.}

\item{...}{Passed to \code{props}.}

\item{label}{Label of the action button.}
}
\description{
Update an action button.
}
