% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-text-area.R, R/input-text.R
\name{lit_text_input_update}
\alias{lit_text_input_update}
\title{Update Text Input}
\usage{
lit_text_input_update(session, selector, ..., value = NULL)

lit_text_input_update(session, selector, ..., value = NULL)
}
\arguments{
\item{session}{A valid shiny session.}

\item{selector}{A selector for the text inputs to update.}

\item{...}{Passed to \code{props}.}

\item{value}{Value of the text input.}
}
\description{
Update text inputs.

Update text inputs.
}
