% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-link.R
\name{litActionLink}
\alias{litActionLink}
\title{Lit Action Link}
\usage{
litActionLink(
  name = "",
  content = "Click me",
  ...,
  id = NULL,
  class = NULL,
  value = NULL,
  callback = NULL,
  send_on_render = FALSE
)
}
\arguments{
\item{name}{Name of the input.}

\item{content}{Content of the button, a character string
or valid \code{shiny::tags}.}

\item{...}{Passed to props.}

\item{id}{Id of the input.}

\item{class}{Any additional classes.}

\item{value}{Value of the input.}

\item{callback}{A JavaScript callback functions that accepts one argument.}

\item{send_on_render}{Whether to send the input value on render.}
}
\description{
Create an action button.
}
\examples{
library(shiny)

ui <- fluidPage(
 litActionLink(
   "btn",
   "First btn"
 ),
 litActionLink(
   "btn",
   "Second btn"
 )
)

server <- function(input, output, session){
 observeEvent(input$btn, {
   print(input$btn)
 })
}

if(interactive())
 shinyApp(ui, server)

}
