% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-color.R
\name{litColorInput}
\alias{litColorInput}
\title{Lit Color Input}
\usage{
litColorInput(
  name = "",
  value = "",
  placeholder = "",
  ...,
  id = NULL,
  class = NULL,
  callback = NULL,
  send_on_render = TRUE
)
}
\arguments{
\item{name}{Name of the input.}

\item{value}{Value of the input.}

\item{placeholder}{Placeholder text.}

\item{...}{Passed to props.}

\item{id}{Id of the input.}

\item{class}{Any additional classes.}

\item{callback}{A JavaScript callback functions that accepts one argument.}

\item{send_on_render}{Whether to send the input value on render}
}
\description{
Create a color input.
}
