% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-datalist.R
\name{litDatalistInput}
\alias{litDatalistInput}
\title{Lit Datalist Input}
\usage{
litDatalistInput(
  name = "",
  choices = list(),
  ...,
  id = NULL,
  value = NULL,
  class = NULL,
  placeholder = NULL,
  callback = NULL,
  send_on_render = TRUE
)
}
\arguments{
\item{name}{Name of the input.}

\item{choices}{List of options.}

\item{...}{Passed to props.}

\item{id}{Id of the input.}

\item{value}{Value of the input.}

\item{class}{Any additional classes.}

\item{placeholder}{Placeholder value.}

\item{callback}{A JavaScript callback functions that accepts one argument.}

\item{send_on_render}{Whether to send the input value on render.}
}
\description{
Create a datalist input.
}
\details{
Can be server-side rendered with \code{\link[=update_datalist_input]{update_datalist_input()}}.
}
\seealso{
\code{\link[=update_datalist_input]{update_datalist_input()}}
}
