% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-select.R
\name{litSelectInput}
\alias{litSelectInput}
\title{Lit Select Input}
\usage{
litSelectInput(
  name = "",
  choices = list(),
  ...,
  id = NULL,
  value = NULL,
  class = NULL
)
}
\arguments{
\item{name}{Name of the input.}

\item{choices}{List of options.}

\item{...}{Passed to props.}

\item{id}{Id of the input.}

\item{value}{Value of the input.}

\item{class}{Any additional classes.}
}
\description{
Create a select input.
}
