% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-selector.R
\name{litSelectorInput}
\alias{litSelectorInput}
\title{Lit Selector Input}
\usage{
litSelectorInput(
  name,
  ...,
  dataset = NULL,
  id = NULL,
  value = NULL,
  class = NULL
)
}
\arguments{
\item{name}{Name of the input.}

\item{...}{Passed to props.}

\item{dataset}{Dataset to retrieve variables.}

\item{id}{Id of the input.}

\item{value}{Value of the input.}

\item{class}{Any additional classes.}
}
\description{
Create a selector input.
}
