% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui.R
\name{litter}
\alias{litter}
\title{Graphical User Interface to the litteR-package}
\usage{
litter(filename = NULL)
}
\arguments{
\item{filename}{name of file containing settings (see vignette for details)}
}
\value{
directory name (invisibly) where all results are stored.
}
\description{
Starts a graphical user interface for analysing litter data.
A Tcl/Tk-dialogue will be started if one or more arguments
are missing.
}
\details{
For details, see our vignette by typing: vignette("litter-manual")
}
