% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-text-area.R
\name{litTextAreaInput}
\alias{litTextAreaInput}
\title{Lit Text Area Input}
\usage{
litTextAreaInput(
  name,
  value = "",
  placeholder = "",
  ...,
  id = NULL,
  return = c("instant", "never")
)
}
\arguments{
\item{name}{Name of the input.}

\item{value}{Value of the input.}

\item{placeholder}{Label of the button.}

\item{...}{Passed to props.}

\item{id}{Id of the input.}

\item{return}{How to return the value to the shiny server
Either \code{instant}, or \code{none}.}
}
\description{
Create a text area input.
}
