% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-toggler.R
\name{litTogglerInput}
\alias{litTogglerInput}
\title{Lit Toggler Input}
\usage{
litTogglerInput(name, display, input, ...)
}
\arguments{
\item{name}{Name of the input.}

\item{display}{The content to display.}

\item{input}{The \code{input} to reveal.}

\item{...}{Passed to props.}
}
\description{
Create an input that is initially hidden but revealed on click.
This hijacks the original input which no longer sends values to the server.
}
