% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{medcouple}
\alias{medcouple}
\alias{medcouple.default}
\title{Medcouple}
\usage{
medcouple(x, ...)

\method{medcouple}{default}(x, ...)
}
\arguments{
\item{x}{numeric vector}

\item{\dots}{further arguments passed to or from other methods.}
}
\value{
medcouple (numeric vector of length 1).
}
\description{
Robust statistic that quantifies the skewness of univariate distributions.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: default method
}}

\note{
This is a naive, but robust en simple implementation. For a more 
efficient implementation see package
\href{https://CRAN.R-project.org/package=robustbase}{robustbase} and the
references section below.
}
\references{
Brys, G., M. Hubert, A. Struyf, 2004. A Robust Measure of Skewness. 
  Journal of Computational and Graphical Statistics 13: 996-1017. 
  \doi{10.1198/106186004X12632}.
}
