% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{p_value}
\alias{p_value}
\title{p-value}
\usage{
p_value(x, ...)
}
\arguments{
\item{x}{object}

\item{\dots}{further arguments passed to or from other methods.}
}
\value{
p-value of a test (numeric vector of length 1).
}
\description{
Extract p-value.
}
