% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{power}
\alias{power}
\title{Statistical Power}
\usage{
power(x, ...)
}
\arguments{
\item{x}{object}

\item{\dots}{further arguments passed to or from other methods.}
}
\value{
power of a test (numeric vector of length 1).
}
\description{
Estimate the statistical power of a test.
}
\references{
\url{https://en.wikipedia.org/wiki/Statistical_power}
}
