% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_litter}
\alias{read_litter}
\title{Read Litter Data}
\usage{
read_litter(file)
}
\arguments{
\item{file}{name of litter file}
}
\value{
\code{tibble} with litter data in long format
}
\description{
Reads litter data from various formats. Currently only the 
OSPAR data snapshot format, a long format, and a wide format
are supported. See the package vignette for more details.
}
