% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_litter_groups}
\alias{read_litter_groups}
\title{Read Litter Groups}
\usage{
read_litter_groups(file)
}
\arguments{
\item{file}{name of litter group file}
}
\value{
\code{tibble} with look-up-table of litter groups
}
\description{
Read the file that links litter codes to litter groups.
See the package vignette for more details.
}
