% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_litter_types}
\alias{read_litter_types}
\title{Read Type Names}
\usage{
read_litter_types(filename, logger = create_logger(level = "INFO"))
}
\arguments{
\item{filename}{name of type file}

\item{logger}{optional logger object (see \code{\link{create_logger}})}
}
\value{
\code{tibble} with look-up-table of type names and group codes
}
\description{
Read the file that links type names to group codes
See the package vignette for more details.
}
