% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_settings}
\alias{read_settings}
\title{Read Settings File}
\usage{
read_settings(filename, logger = create_logger(level = "INFO"))
}
\arguments{
\item{filename}{name of litter file}

\item{logger}{optional logger object (see \code{\link{create_logger}})}
}
\value{
validated settings file
}
\description{
Reads settings file. See tutorial for its format.
}
