% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{recdf}
\alias{recdf}
\title{Sample From an ECDF}
\usage{
recdf(x, n)
}
\arguments{
\item{x}{numeric vector}

\item{n}{number of draws}
}
\value{
vector of \code{n} elements of the same type as \code{x}
}
\description{
Type stable implementation of an Empirical Cumulative Distribution
Function (ECDF) sampler.
}
\examples{
recdf(1:5, 10)
}
\seealso{
\code{\link{ecdf}}
}
