% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{rmad}
\alias{rmad}
\title{Relative Median Absolute Deviation}
\usage{
rmad(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector}

\item{na.rm}{logical. Should missing values be removed?}
}
\value{
Relative median absolute deviation (numeric vector of length 1).
}
\description{
This is the Median Absolute Deviation divided by the median and
is similar to the coefficient of variation.
}
\references{
\url{https://en.wikipedia.org/wiki/Median_absolute_deviation}
}
