% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{roll}
\alias{roll}
\title{Rolling Statistics}
\usage{
roll(x, w = 3, fun = mean)
}
\arguments{
\item{x}{numeric vector (time-series)}

\item{w}{width of moving window}

\item{fun}{function to be applied}
}
\value{
vector of length length(x)-w
}
\description{
Applies function \code{fun} within a rolling (moving) window
of size \code{w} to vector numeric vector \code{x}.
}
