% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sequenize.integer}
\alias{sequenize.integer}
\title{Sequenize Integer Sequence}
\usage{
\method{sequenize}{integer}(x, ...)
}
\arguments{
\item{x}{vector of integers.}

\item{\dots}{further arguments passed to or from other methods.}
}
\value{
object of class \code{sequenized}
}
\description{
Compression of integer sequences to 'start-end' notation.
For instance \code{c(1:5, 8:9)} becomes \code{"1-5, 8-9"}.
}
\note{
The elements of \code{x} should be unique and in ascending order.
}
\examples{
toString(sequenize(c(1:4, 8:9)))
}
\seealso{
\code{\link{toString}}
}
