% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sequenize}
\alias{sequenize}
\title{Sequenize Objects}
\usage{
sequenize(x, ...)
}
\arguments{
\item{x}{object to sequenize}

\item{\dots}{further arguments passed to or from other methods.}
}
\description{
Generic function for sequenizing objects
}
\seealso{
\code{\link{sequenize.integer}}
}
