% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{slope}
\alias{slope}
\title{Slope}
\usage{
slope(x, ...)
}
\arguments{
\item{x}{object}

\item{\dots}{further arguments passed to or from other methods.}
}
\value{
estimate of the slope (numeric vector of length 1).
}
\description{
Extract slope.
}
