% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/els-utils.R
\name{tag2}
\alias{tag2}
\title{Custom Tag}
\usage{
tag2(.name, ..., .script = NULL)
}
\arguments{
\item{.name}{Name of the tag.}

\item{...}{Tag attributes (named).}
}
\description{
Create a custom tag.
}
\keyword{internal}
