% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{test_statistic}
\alias{test_statistic}
\title{Test Statistic}
\usage{
test_statistic(x, ...)
}
\arguments{
\item{x}{object}

\item{\dots}{further arguments passed to or from other methods.}
}
\value{
test statistic of a test (numeric vector of length 1).
}
\description{
Extract test_statistic.
}
\seealso{
\code{\link{test_statistic.wilcoxon}}, \code{\link{test_statistic.mann_kendall}}
}
