% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{top}
\alias{top}
\title{Select Most Abundant Litter Types}
\usage{
top(type, count, fraction = 0.8)
}
\arguments{
\item{type}{character vector of litter types}

\item{count}{counts for each litter type}

\item{fraction}{fraction of most abundant types (numeric value in [0, 1])}
}
\value{
Fraction of most abundant litter types in descending order
        (numeric vector of length 1).
}
\description{
Select the \code{fraction} most abundant litter types.
}
