% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_input}
\alias{update_input}
\title{Update input}
\usage{
update_input(
  ...,
  name = NULL,
  id = NULL,
  selector = NULL,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{...}{Props to update, e.g.: \code{value}.}

\item{name}{Name of the input.}

\item{id}{Id of the input.}

\item{selector}{A CSS selector.}

\item{session}{A Shiny session.}
}
\description{
Update an input
}
\details{
Set either of \code{name}, \code{id}, or \code{selector}.
}
