% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{validate}
\alias{validate}
\alias{validate.ospar}
\alias{validate.long}
\alias{validate.wide}
\alias{validate.litter_group}
\title{Validation of Litter File Formats}
\usage{
validate(x, ...)

\method{validate}{ospar}(x, ...)

\method{validate}{long}(x, ...)

\method{validate}{wide}(x, ...)

\method{validate}{litter_group}(x, ...)
}
\arguments{
\item{x}{object to validate}

\item{\dots}{further arguments passed to or from other methods.}
}
\value{
validated object of class \code{ospar}

validated object of class \code{long}

validated object of class \code{wide}

validated object of class \code{litter_group}
}
\description{
Generic function for validation of file formats.
}
\section{Methods (by class)}{
\itemize{
\item \code{ospar}: validate OSPAR data.

\item \code{long}: validate long format data.

\item \code{wide}: validate wide format data.

\item \code{litter_group}: validate litter_group file
}}

