% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{enumerate}
\alias{enumerate}
\alias{enumerate.numeric}
\title{Enumerate Objects}
\usage{
enumerate(x, ...)

\method{enumerate}{numeric}(x, ...)
}
\arguments{
\item{x}{object to enumerate}

\item{\dots}{further arguments passed to or from other methods.}
}
\description{
Generic function for enumerating objects
}
\section{Methods (by class)}{
\itemize{
\item \code{numeric}: enumerate \code{numeric} vector.
}}

\seealso{
\code{\link{enumerate.character}}
}
