% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-button.R
\name{lit_action_button_update}
\alias{lit_action_button_update}
\title{Update Action Button}
\usage{
lit_action_button_update(session, selector, ..., content = NULL)
}
\arguments{
\item{session}{A valid shiny session.}

\item{selector}{A selector for the buttons to update.}

\item{...}{Passed to props.}

\item{content}{Content of the button, a character string
or valid \code{shiny::tags}.}
}
\description{
Update a action buttons.
}
\examples{
library(shiny)

ui <- fluidPage(
	litActionButton(
		"btn",
		"First btn"
	),
	litActionButton(
		"btn",
		"Second btn"
	),
	litActionButton(
		"update",
		"Parent"
	)
)

server <- function(input, output, session){
	observeEvent(input$btn, {
		print(input$btn)
	})

	observeEvent(input$update, {
		lit_action_button_update(
			session,
			selector = select_inputs("btn"),
			label = "LABEL",
			x = 1
		)
	})
}

if(interactive())
	shinyApp(ui, server)

}
