% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-button.R
\name{litActionButton}
\alias{litActionButton}
\title{Lit Action Button}
\usage{
litActionButton(name, label = "Click me", ..., id = NULL, class = NULL)
}
\arguments{
\item{name}{Name of the input.}

\item{label}{Label of the button.}

\item{...}{Passed to props.}

\item{id}{Id of the input.}

\item{class}{Class of the button.}
}
\description{
Create an action button.
}
