% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-button.R
\name{litActionButton}
\alias{litActionButton}
\title{Lit Action Button}
\usage{
litActionButton(
  name = "",
  content = "Click me",
  ...,
  id = NULL,
  class = NULL,
  value = NULL
)
}
\arguments{
\item{name}{Name of the input.}

\item{content}{Content of the button, a character string
or valid \code{shiny::tags}.}

\item{...}{Passed to props.}

\item{id}{Id of the input.}

\item{class}{Any additional classes.}

\item{value}{Value of the input.}
}
\description{
Create an action button.
}
\examples{
library(shiny)

ui <- fluidPage(
	litActionButton(
		"btn",
		"First btn"
	),
	litActionButton(
		"btn",
		"Second btn"
	)
)

server <- function(input, output, session){
	observeEvent(input$btn, {
		print(input$btn)
	})
}

if(interactive())
	shinyApp(ui, server)

}
