% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-button.R
\name{litActionButton}
\alias{litActionButton}
\title{Lit Action Button}
\usage{
litActionButton(
  name,
  content = "Click me",
  ...,
  id = NULL,
  callback = NULL,
  color = lit_default_color(),
  size = c("standard", "small", "large")
)
}
\arguments{
\item{name}{Name of the input.}

\item{content}{Content of the button, a character string
or valid \code{shiny::tags}.}

\item{...}{Passed to props.}

\item{id}{Id of the input.}

\item{callback}{Callback function to run on click
if \code{NULL} then the value is sent to the server.
Otherwise a JavaScript callback function that accepts
the instance of the button class (e.g.: \code{this}).}

\item{color}{Color of button, e.g.: \code{outline-primary}.}

\item{size}{Size of the button.}
}
\description{
Create an action button.
}
\examples{
library(shiny)

ui <- fluidPage(
	litActionButton(
		"btn",
		"First btn"
	),
	litActionButton(
		"btn",
		"Second btn"
	)
)

server <- function(input, output, session){
	observeEvent(input$btn, {
		print(input$btn)
	})
}

if(interactive())
	shinyApp(ui, server)

}
