% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-checkboxes.R
\name{litCheckboxesInput}
\alias{litCheckboxesInput}
\title{Lit Checkboxes Input}
\usage{
litCheckboxesInput(
  name = "",
  choices = list(),
  ...,
  id = NULL,
  value = list(),
  class = NULL,
  callback = NULL
)
}
\arguments{
\item{name}{Name of the input.}

\item{choices}{List of options.}

\item{...}{Passed to props.}

\item{id}{Id of the input.}

\item{value}{\code{choices} that are checked.}

\item{class}{Any additional classes.}

\item{callback}{A JavaScript callback functions that accepts one argument.}
}
\description{
Create checkboxes input.
}
