% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-color.R
\name{litColorInput}
\alias{litColorInput}
\title{Lit Color Input}
\usage{
litColorInput(name, value = "", placeholder = "", ..., id = NULL, class = NULL)
}
\arguments{
\item{name}{Name of the input.}

\item{value}{Value of the input.}

\item{placeholder}{Placeholder text.}

\item{...}{Passed to props.}

\item{id}{Id of the input.}

\item{class}{Any additional classes.}
}
\description{
Create a color input.
}
