% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-filter.R
\name{litFilterInput}
\alias{litFilterInput}
\title{Lit Filter Input}
\usage{
litFilterInput(
  name,
  ...,
  dataset = NULL,
  id = NULL,
  value = NULL,
  class = NULL
)
}
\arguments{
\item{name}{Name of the input.}

\item{...}{Passed to props.}

\item{dataset}{Dataset to retrieve variables.}

\item{id}{Id of the input.}

\item{value}{Value of the input.}

\item{class}{Any additional classes.}
}
\description{
Create a filter input.
}
