% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-radios.R
\name{litRadioInput}
\alias{litRadioInput}
\title{Lit Radio Input}
\usage{
litRadioInput(
  name = "",
  choices = list(),
  ...,
  id = NULL,
  value = NULL,
  class = NULL,
  callback = NULL,
  send_on_render = TRUE
)
}
\arguments{
\item{name}{Name of the input.}

\item{choices}{List of options.}

\item{...}{Passed to props.}

\item{id}{Id of the input.}

\item{value}{\code{choices} that are checked.}

\item{class}{Any additional classes.}

\item{callback}{A JavaScript callback functions that accepts one argument.}

\item{send_on_render}{Whether to send the input value on render.}
}
\description{
Create radio input.
}
