% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-range.R
\name{litRangeInput}
\alias{litRangeInput}
\title{Lit Range Input}
\usage{
litRangeInput(
  name = "",
  ...,
  id = NULL,
  min = 0L,
  max = 10L,
  step = NULL,
  value = NULL,
  class = NULL,
  callback = ""
)
}
\arguments{
\item{name}{Name of the input.}

\item{...}{Passed to props.}

\item{id}{Id of the input.}

\item{min, max, value}{Minimum, maximum, and initial value
of the range.}

\item{step}{Incremental steps the range must take.}

\item{class}{Any additional classes.}

\item{callback}{A JavaScript callback functions that accepts one argument.}
}
\description{
Create a range input.
}
\examples{
library(shiny)

ui <- fluidPage(
	 litRangeInput("range")
)

server <- function(input, output, session){
  observeEvent(input$range, {
    print(input$range)
  })
}

if(interactive())
  shinyApp(ui, server)

}
