% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-select.R
\name{litSelectInput}
\alias{litSelectInput}
\title{Lit Select Input}
\usage{
litSelectInput(name, options = list(), ..., id = NULL, class = NULL)
}
\arguments{
\item{name}{Name of the input.}

\item{options}{List of options (\code{name} & \code{value}).}

\item{...}{Passed to props.}

\item{id}{Id of the input.}

\item{class}{Class of the button.}
}
\description{
Create a select input.
}
