% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-switch.R
\name{litSwitchInput}
\alias{litSwitchInput}
\title{Lit Switch Input}
\usage{
litSwitchInput(
  name = "",
  value = TRUE,
  label = "",
  ...,
  id = NULL,
  class = NULL,
  callback = ""
)
}
\arguments{
\item{name}{Name of the input.}

\item{value}{Value of the input.}

\item{label}{Label of the button.}

\item{...}{Passed to props.}

\item{id}{Id of the input.}

\item{class}{Any additional classes.}

\item{callback}{A JavaScript callback functions that accepts one argument.}
}
\description{
Create a switch input.
}
