% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui.R
\name{litter}
\alias{litter}
\title{Graphical User Interface to the litteR-package}
\usage{
litter(file_settings = NULL, file_input = NULL, dir_output = NULL)
}
\arguments{
\item{file_settings}{file containing settings}

\item{file_input}{file containing litter data}

\item{dir_output}{directory to store output}
}
\value{
An HTML-document in which all the litter analysis results 
(tables, figures, explanotory text) are reported.
}
\description{
Starts a graphical user interface for analysing litter data. 
A Tcl/Tk-dialogue will be started if one or more arguments 
are missing.
}
