% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui.R
\name{litter}
\alias{litter}
\title{Graphical User Interface to the litteR-package}
\usage{
litter(file = NULL)
}
\arguments{
\item{file}{file containing litter data (see vignette for details)}
}
\value{
An HTML-document in which all the litter analysis results 
(tables, figures, explanotory text) are reported.
}
\description{
Starts a graphical user interface for analysing litter data. 
A Tcl/Tk-dialogue will be started if one or more arguments 
are missing.
}
