% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-text-area.R
\name{litTextAreaInput}
\alias{litTextAreaInput}
\title{Lit Text Area Input}
\usage{
litTextAreaInput(
  name = "",
  value = "",
  placeholder = "",
  ...,
  id = NULL,
  class = NULL
)
}
\arguments{
\item{name}{Name of the input.}

\item{value}{Value of the input.}

\item{placeholder}{Placeholder text.}

\item{...}{Passed to props.}

\item{id}{Id of the input.}

\item{class}{Any additional classes.}
}
\description{
Create a text area input.
}
