% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-text.R
\name{litTextInput}
\alias{litTextInput}
\title{Lit Text Input}
\usage{
litTextInput(
  name,
  value = "",
  placeholder = "",
  ...,
  id = NULL,
  return = c("enter", "instant", "never")
)
}
\arguments{
\item{name}{Name of the input.}

\item{value}{Value of the input.}

\item{placeholder}{Label of the button.}

\item{...}{Passed to props.}

\item{id}{Id of the input.}

\item{return}{How to return the value to the shiny server
Either \code{instant} (not recommended), \code{none} or \code{enter}, the
latter sends it on enter.}
}
\description{
Create a text input.
}
