% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-text-label.R
\name{litTextLabelInput}
\alias{litTextLabelInput}
\title{Lit Text Label Input}
\usage{
litTextLabelInput(
  name = "",
  value = "",
  label = "",
  placeholder = label,
  ...,
  id = NULL,
  class = NULL
)
}
\arguments{
\item{name}{Name of the input.}

\item{value}{Value of the input.}

\item{label}{Label of the input.}

\item{placeholder}{Placeholder text.}

\item{...}{Passed to props.}

\item{id}{Id of the input.}

\item{class}{Any additional classes.}
}
\description{
Create a text input with a floating label.
}
