% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-toggler.R
\name{litTogglerInput}
\alias{litTogglerInput}
\title{Lit Toggler Input}
\usage{
litTogglerInput(name, display, input, ..., callback = NULL, restore = TRUE)
}
\arguments{
\item{name}{Name of the input.}

\item{display}{The content to display.}

\item{input}{The \code{input} to reveal.}

\item{...}{Passed to props.}

\item{callback}{A JavaScript callback functions that accepts one argument.}

\item{restore}{Whether to set the input \code{value} to the \code{display} text and vice versa.}
}
\description{
Create an input that is initially hidden but revealed on click.
This hijacks the original input which no longer sends values to the server.
}
\details{
This can be updated with the \code{feedback} argument in \code{\link[=update_input]{update_input()}}
as well as \code{accept}, a boolean to indicate whether to accept the value set.
}
