% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_litter}
\alias{read_litter}
\title{Read Litter Data}
\usage{
read_litter(filename, logger = create_logger(level = "INFO"), type_names)
}
\arguments{
\item{filename}{name of litter file}

\item{logger}{optional logger object (see \code{\link{create_logger}})}

\item{type_names}{character vector of allowed type_names}
}
\value{
\code{tibble} with litter data in long format
}
\description{
Reads litter data from various formats. Currently only the
OSPAR data snapshot format, and a wide format
are supported. See the package vignette for more details.
}
